add Makefile && fix errors 	- fix complie error 	- fix Libftest error 	- add 42Header

+++ Libft/ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: siseo <siseo@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 17:18:13 by siseo             #+#    #+#             */
-/*   Updated: 2022/03/21 17:03:38 by siseo            ###   ########.fr       */
+/*   Updated: 2022/03/25 16:53:26 by siseo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	size_t	i;
+	int	i;
 
 	if ((size_t)(dst - src) >= len)
 		ft_memcpy(dst, src, len);
 	else
 	{
-		i = len - 1;
+		i = (int)(len - 1);
 		while (i >= 0)
-		{
+		{	
 			((unsigned char *)(dst))[i] = ((const unsigned char *)(src))[i];
 			i--;
 		}
